Extract tests specific to Remix v2 flags
